fix(tests): slug-as-project-ref fallout from 1ce031eae path-column drop
After `1ce031eae` removed `projects.path`, `ProjectRepository::resolve`
only accepts UUIDs or `owner/repo` slugs. Callers that were still
handing it a filesystem path silently returned None, producing
"project not found" errors and downstream assertion failures across
~30 tests. This batch cuts those callers over to the slug/UUID
contract and patches two production handlers that still treated
`p.project` as a path:

- `memory_tools/move_ops`: derive canonical project path from the
  resolved project row (mirrors `create_note`), so moved notes land
  under `project_dir(owner, repo)/.djinn/...` where `propose_adr_list`
  actually scans.
- `proposal_tools` show/accept/reject: resolve via `ProjectRepository`
  instead of `PathBuf::from(&p.project)`. `propose_adr_list` already
  did this; the other three arms were drift.
- `agent/extension/handlers`: add `project_ref` (slug when resolved,
  worktree path as fallback) and route memory/agent MCP calls through
  that; also try `args["project"]` when the worktree path itself
  doesn't resolve. Removes the dispatcher short-circuit for
  `code_graph` so the stub-surfaced "not available" error still
  reaches agent tests with empty DBs.
- `verification/environment::resolve_project_id_for_path`: walk
  ancestors and match `{owner}/{repo}` on each window, so workers
  operating on `{clone}/.djinn/worktrees/<id>` still find their
  project row (the last-two-components parse was too strict).

Schema-side: `ChangedRange` + the `code_graph` / `pr_review_context`
numeric Params fields flip from `u32`/`usize` to `i64` so schemars
stops emitting nonstandard `uint`/`uint32` formats (which
`mcp_tools_list_schemas_do_not_use_nonstandard_uint_or_nullable_without_type`
enforces). Cast at the bridge boundary where the graph layer still
takes `u32`.

Tests updated to pass `project.slug()` (or a fresh unique
owner/repo) instead of `tmp.path()`, with a local `PathCleanupGuard`
in the few cases that need to write into the real
`project_dir(owner, repo)` location. `create_test_app_with_project`
now returns the slug directly — contract tests always called it as
an MCP arg, never as a path.

Two snapshots accepted unchanged in shape, only regenerated:
`architect_tools_section_snapshot` (reflects the schema rename +
new code_graph ops) and `mcp_tools_schema` (chat sessions'
`project_id` is nullable since ADR-054).

Also bundles the pre-existing WIP chat session-title test scaffolding
(`chat/tests/sessions_endpoints.rs`, the `session_title_sse_payload_shape`
case in `sse.rs`, and the mod-rs wiring) which was already in the
working tree before this batch.

Covered tests moved from FAIL → PASS:
- djinn-control-plane: 245 tests across lib + 13 integration binaries
- djinn-server: 145 lib tests (including mcp_contract_tests::memory_tools)
- djinn-agent: 607 lib tests (verification, extension, prompts snapshot)

One pre-existing unrelated failure remains:
`djinn-agent::process::tests::spawned_process_uses_different_pgid`
(environmental — `ps -o pgid=` returns empty on this box, not touched
by this refactor).

+++ server/crates/djinn-control-plane/src/bridge.rs
@@ -222,10 +222,10 @@ pub struct ChangedRange {
     /// Repository-relative path of the file the hunk lives in.
     pub file: String,
     /// Inclusive 1-indexed first line of the hunk.
-    pub start_line: u32,
+    pub start_line: i64,
     /// Inclusive 1-indexed last line of the hunk. Defaults to `start_line`
     /// when the caller passed a single-line hunk.
-    pub end_line: Option<u32>,
+    pub end_line: Option<i64>,
 }
 
 /// A single symbol (or file) whose definition range encloses a queried
